refactor: pisahkan HotelRoomBooking menjadi beberapa kelas berdasarkan tanggung jawab
