+++ 08_Assesment_Tugas_Besar_Pengganti_UTS/HotelRoomBooking.java
@@ -3,162 +3,20 @@ import java.util.Date;
 
 public class HotelRoomBooking {
 
-    private String namaPemesan;
-    private String nomorKTP;
-    private String nomorTelepon;
-    private String email;
-    private String jenisKelamin;
-    private String jenisKamar;
-    private int jumlahTamu;
-    private int jumlahMalam;
-    private double hargaPerMalam;
-    private String kodeVoucher;
-    private boolean sudahDibayar;
-    private Date tanggalCheckin;
-    private Date tanggalCheckout;
-    private boolean statusAktif;
+    public static void main(String[] args) {
+        // Membuat objek Pemesan
+        Pemesan pemesan = new Pemesan("Aorinka Anendya Chazanah", "1234567890123456", "081234567890", "aorinka@gmail.com", "Wanita");
 
-    public HotelRoomBooking(String namaPemesan, String nomorKTP, String nomorTelepon, String email, String jenisKelamin,
-            String jenisKamar, int jumlahTamu, int jumlahMalam, double hargaPerMalam,
-            Date tanggalCheckin, Date tanggalCheckout, boolean statusAktif,
-            String kodeVoucher, boolean sudahDibayar) {
+        // Membuat objek Kamar
+        Kamar kamar = new Kamar("Deluxe", 3, 5, 500000, new Date(), new Date());
 
-        this.namaPemesan = namaPemesan;
-        this.nomorKTP = nomorKTP;
-        this.nomorTelepon = nomorTelepon;
-        this.email = email;
-        this.jenisKelamin = jenisKelamin;
-        this.jenisKamar = jenisKamar;
-        this.jumlahTamu = jumlahTamu;
-        this.jumlahMalam = jumlahMalam;
-        this.hargaPerMalam = hargaPerMalam;
-        this.tanggalCheckin = tanggalCheckin;
-        this.tanggalCheckout = tanggalCheckout;
-        this.statusAktif = statusAktif;
-        this.kodeVoucher = kodeVoucher;
-        this.sudahDibayar = sudahDibayar;
-    }
+        // Membuat objek Pembayaran
+        Pembayaran pembayaran = new Pembayaran("VOUCHER123", false, true);
 
-    public void cetakDetailPemesanan() {
-        System.out.println("===== DETAIL PEMESANAN KAMAR =====");
-        cetakInformasiPemesan();
-        cetakInformasiKamar();
-        cetakTanggalDanStatus();
-        cetakPembayaran();
-        System.out.println("===================================");
-    }
+        // Membuat objek Pemesanan dengan menggabungkan Pemesan, Kamar, dan Pembayaran
+        Pemesanan pemesanan = new Pemesanan(pemesan, kamar, pembayaran);
 
-    private void cetakInformasiPemesan() {
-        System.out.println("Nama Pemesan : " + namaPemesan);
-        System.out.println("Jenis Kelamin: " + jenisKelamin);
-        System.out.println("No. KTP      : " + nomorKTP);
-        System.out.println("Telepon      : " + nomorTelepon);
-        System.out.println("Email        : " + email);
-    }
-
-    private void cetakInformasiKamar() {
-        System.out.println("Jenis Kamar  : " + jenisKamar);
-        System.out.println("Jumlah Tamu  : " + jumlahTamu);
-        System.out.println("Jumlah Malam : " + jumlahMalam);
-        System.out.println("Harga/Malam  : " + hargaPerMalam);
-    }
-
-    private void cetakTanggalDanStatus() {
-        System.out.println("Check-in     : " + tanggalCheckin);
-        System.out.println("Check-out    : " + tanggalCheckout);
-        System.out.println("Status Aktif : " + statusAktif);
-    }
-
-    private void cetakPembayaran() {
-        System.out.println("Voucher      : " + kodeVoucher);
-        System.out.println("Sudah Dibayar: " + sudahDibayar);
-        System.out.println("Total Biaya  : Rp " + hitungTotalBiaya());
-        System.out.println("Tipe Tamu    : " + klasifikasiTamu());
-    }
-
-    public void redeemVoucher(String kode) {
-    }
-
-    public void kirimEmailKonfirmasi() {
-    }
-
-    public double hitungTotalBiaya() {
-        double total = hargaPerMalam * jumlahMalam;
-        if (jumlahTamu > 2) {
-            total += (jumlahTamu - 2) * 100000;
-        }
-        if (kodeVoucher != null && kodeVoucher.length() > 3) {
-            total -= 50000;
-        }
-        if (!statusAktif) {
-            total = 0;
-        }
-        return total;
-    }
-
-    public String klasifikasiTamu() {
-        if (jumlahTamu == 1) {
-            return "Individu";
-        } else if (jumlahTamu == 2) {
-            return "Pasangan";
-        } else {
-            return "Keluarga";
-        }
-    }
-
-    public String getNamaPemesan() {
-        return namaPemesan;
-    }
-
-    public void setNamaPemesan(String namaPemesan) {
-        this.namaPemesan = namaPemesan;
-    }
-
-    public String getNomorKTP() {
-        return nomorKTP;
-    }
-
-    public void setNomorKTP(String nomorKTP) {
-        this.nomorKTP = nomorKTP;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getJenisKamar() {
-        return jenisKamar;
-    }
-
-    public void setJenisKamar(String jenisKamar) {
-        this.jenisKamar = jenisKamar;
-    }
-
-    public int getJumlahMalam() {
-        return jumlahMalam;
-    }
-
-    public void setJumlahMalam(int jumlahMalam) {
-        this.jumlahMalam = jumlahMalam;
-    }
-
-    public double getHargaPerMalam() {
-        return hargaPerMalam;
-    }
-
-    public void setHargaPerMalam(double hargaPerMalam) {
-        this.hargaPerMalam = hargaPerMalam;
-    }
-
-    public boolean isSudahDibayar() {
-        return sudahDibayar;
-    }
-
-    public void setSudahDibayar(boolean sudahDibayar) {
-        this.sudahDibayar = sudahDibayar;
+        // Cetak detail pemesanan
+        pemesanan.cetakDetailPemesanan();
     }
 }
