+++ 08_Assesment_Tugas_Besar_Pengganti_UTS/Kamar.java
@@ -1,0 +1,38 @@
+
+import java.util.Date;
+
+public class Kamar {
+
+    private String jenisKamar;
+    private int jumlahTamu;
+    private int jumlahMalam;
+    private double hargaPerMalam;
+    private Date tanggalCheckin;
+    private Date tanggalCheckout;
+
+    public Kamar(String jenisKamar, int jumlahTamu, int jumlahMalam, double hargaPerMalam, Date tanggalCheckin, Date tanggalCheckout) {
+        this.jenisKamar = jenisKamar;
+        this.jumlahTamu = jumlahTamu;
+        this.jumlahMalam = jumlahMalam;
+        this.hargaPerMalam = hargaPerMalam;
+        this.tanggalCheckin = tanggalCheckin;
+        this.tanggalCheckout = tanggalCheckout;
+    }
+
+    public void printInformasi() {
+        System.out.println("Jenis Kamar  : " + jenisKamar);
+        System.out.println("Jumlah Tamu  : " + jumlahTamu);
+        System.out.println("Jumlah Malam : " + jumlahMalam);
+        System.out.println("Harga/Malam  : " + hargaPerMalam);
+        System.out.println("Check-in     : " + tanggalCheckin);
+        System.out.println("Check-out    : " + tanggalCheckout);
+    }
+
+    public double hitungTotalBiaya() {
+        double total = hargaPerMalam * jumlahMalam;
+        if (jumlahTamu > 2) {
+            total += (jumlahTamu - 2) * 100000;
+        }
+        return total;
+    }
+}
