+++ 08_Assesment_Tugas_Besar_Pengganti_UTS/Pemesanan.java
@@ -1,0 +1,22 @@
+
+public class Pemesanan {
+
+    private Pemesan pemesan;
+    private Kamar kamar;
+    private Pembayaran pembayaran;
+
+    public Pemesanan(Pemesan pemesan, Kamar kamar, Pembayaran pembayaran) {
+        this.pemesan = pemesan;
+        this.kamar = kamar;
+        this.pembayaran = pembayaran;
+    }
+
+    public void cetakDetailPemesanan() {
+        pemesan.printInformasi();
+        kamar.printInformasi();
+        pembayaran.printInformasi();
+        double totalBiaya = kamar.hitungTotalBiaya();
+        totalBiaya = pembayaran.hitungTotalBiaya(totalBiaya);
+        System.out.println("Total Biaya  : Rp " + totalBiaya);
+    }
+}
